[C] Refactor Review page to use server
