+++ components/questions/Review/Select/index.tsx
@@ -7,21 +7,22 @@ import { Option } from "@/components/shapes/option";
 import { getLabelByValue } from "@/components/questions/utils";
 import { BaseReviewProps } from "@/types/questions";
 import { SelectInput } from "@/types/answers";
+import useAnswer from "@/hooks/useAnswer";
 
-interface SelectProps extends BaseReviewProps<SelectInput> {
+interface SelectProps extends BaseReviewProps {
   options: Array<Option>;
   questionText: string;
 }
 
 const SelectReview: FunctionComponent<SelectProps> = ({
-  value,
+  id,
   options = [],
   number,
   questionText,
 }) => {
   const { t } = useTranslation();
-
-  const label = getLabelByValue(options, value);
+  const { answer } = useAnswer<SelectInput>(id);
+  const label = getLabelByValue(options, answer);
 
   return (
     <Styled.ReviewListItem value={number}>
