+++ components/questions/Review/Text/index.tsx
@@ -5,22 +5,24 @@ import { useTranslation } from "react-i18next";
 import * as Styled from "../styles";
 import { BaseReviewProps } from "@/types/questions";
 import { TextInput } from "@/types/answers";
+import useAnswer from "@/hooks/useAnswer";
 
 interface TextProps extends BaseReviewProps<TextInput> {
   questionText: string;
 }
 
 const TextReview: FunctionComponent<TextProps> = ({
-  value,
+  id,
   number,
   questionText,
 }) => {
   const { t } = useTranslation();
+  const { answer } = useAnswer<TextInput>(id);
 
   return (
     <Styled.ReviewListItem value={number}>
       {questionText}
-      <Styled.Text>{value || t("review.no_answer")}</Styled.Text>
+      <Styled.Text>{answer || t("review.no_answer")}</Styled.Text>
     </Styled.ReviewListItem>
   );
 };
