+++ components/questions/Review/index.ts
@@ -2,3 +2,4 @@ export { default as Text } from "./Text";
 export { default as Select } from "./Select";
 export { default as Widget } from "./Widget";
 export { default as Inline } from "./Inline";
+export { default as Tabular } from "./Tabular";
