+++ components/questions/ReviewList/index.tsx
@@ -1,25 +1,17 @@
-"use client";
 import { FunctionComponent } from "react";
 import ReviewFactory from "@/components/factories/ReviewFactory";
-import { StoredQuestion } from "@/contexts/Questions";
+import { StoredQuestion } from "@/helpers/questions";
 import * as Styled from "./styles";
-import { Answers } from "@/types/answers";
 
 const ReviewList: FunctionComponent<{
-  answers: Answers;
   questions: Array<StoredQuestion>;
-}> = ({ answers, questions }) => {
+}> = ({ questions }) => {
   return (
     <Styled.ReviewList>
       {questions &&
         questions.map(({ id, answerType: type, ...config }, i) => {
           return (
-            <ReviewFactory
-              key={id}
-              number={i + 1}
-              value={answers[id]?.data}
-              {...{ id, type, config }}
-            />
+            <ReviewFactory key={id} number={i + 1} {...{ id, type, config }} />
           );
         })}
     </Styled.ReviewList>
