+++ gql/public-schema/gql.ts
@@ -25,7 +25,7 @@ const documents = {
     types.InvestigationIdDocument,
   "\n  query InvestigationPage($site: [String], $uri: [String]) {\n    entry(site: $site, uri: $uri) {\n      __typename\n      ...InvestigationLandingPageTemplate\n    }\n  }\n":
     types.InvestigationPageDocument,
-  "\n  query ReviewPage($site: [String], $uri: [String]) {\n    entry(site: $site, uri: $uri) {\n      ... on investigations_investigationParent_Entry {\n        title\n      }\n    }\n  }\n":
+  "\n  query ReviewPage($site: [String], $uri: [String]) {\n    entry(site: $site, uri: $uri) {\n      ... on investigations_investigationParent_Entry {\n        title\n        children {\n          ... on investigations_default_Entry {\n            contentBlocks {\n              ...QuestionsBlock\n              ... on contentBlocks_twoColumnContainer_BlockType {\n                columns: children {\n                  ... on contentBlocks_colLeft_BlockType {\n                    children {\n                      ...QuestionsBlock\n                      ... on contentBlocks_group_BlockType {\n                        group: children {\n                          ...QuestionsBlock\n                        }\n                      }\n                    }\n                  }\n                  ... on contentBlocks_colRight_BlockType {\n                    children {\n                      ...QuestionsBlock\n                      ... on contentBlocks_group_BlockType {\n                        group: children {\n                          ... on contentBlocks_questionBlock_BlockType {\n                            questionEntries {\n                              ...QuestionEntry\n                            }\n                          }\n                        }\n                      }\n                    }\n                  }\n                }\n              }\n              ... on contentBlocks_group_BlockType {\n                group: children {\n                  ...QuestionsBlock\n                }\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n":
     types.ReviewPageDocument,
   '\n  query GlobalsQuery($site: [String]) {\n    siteInfo: globalSet(site: $site, handle: "siteInfo") {\n      ... on siteInfo_GlobalSet {\n        language\n        name\n        handle\n        siteTitle\n        siteDescription\n      }\n    }\n    menuContent: globalSet(site: $site, handle: "menuContent") {\n      ... on menuContent_GlobalSet {\n        helpUrl\n      }\n    }\n  }\n':
     types.GlobalsQueryDocument,
@@ -179,8 +179,8 @@ export function graphql(
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(
-  source: "\n  query ReviewPage($site: [String], $uri: [String]) {\n    entry(site: $site, uri: $uri) {\n      ... on investigations_investigationParent_Entry {\n        title\n      }\n    }\n  }\n"
-): (typeof documents)["\n  query ReviewPage($site: [String], $uri: [String]) {\n    entry(site: $site, uri: $uri) {\n      ... on investigations_investigationParent_Entry {\n        title\n      }\n    }\n  }\n"];
+  source: "\n  query ReviewPage($site: [String], $uri: [String]) {\n    entry(site: $site, uri: $uri) {\n      ... on investigations_investigationParent_Entry {\n        title\n        children {\n          ... on investigations_default_Entry {\n            contentBlocks {\n              ...QuestionsBlock\n              ... on contentBlocks_twoColumnContainer_BlockType {\n                columns: children {\n                  ... on contentBlocks_colLeft_BlockType {\n                    children {\n                      ...QuestionsBlock\n                      ... on contentBlocks_group_BlockType {\n                        group: children {\n                          ...QuestionsBlock\n                        }\n                      }\n                    }\n                  }\n                  ... on contentBlocks_colRight_BlockType {\n                    children {\n                      ...QuestionsBlock\n                      ... on contentBlocks_group_BlockType {\n                        group: children {\n                          ... on contentBlocks_questionBlock_BlockType {\n                            questionEntries {\n                              ...QuestionEntry\n                            }\n                          }\n                        }\n                      }\n                    }\n                  }\n                }\n              }\n              ... on contentBlocks_group_BlockType {\n                group: children {\n                  ...QuestionsBlock\n                }\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n"
+): (typeof documents)["\n  query ReviewPage($site: [String], $uri: [String]) {\n    entry(site: $site, uri: $uri) {\n      ... on investigations_investigationParent_Entry {\n        title\n        children {\n          ... on investigations_default_Entry {\n            contentBlocks {\n              ...QuestionsBlock\n              ... on contentBlocks_twoColumnContainer_BlockType {\n                columns: children {\n                  ... on contentBlocks_colLeft_BlockType {\n                    children {\n                      ...QuestionsBlock\n                      ... on contentBlocks_group_BlockType {\n                        group: children {\n                          ...QuestionsBlock\n                        }\n                      }\n                    }\n                  }\n                  ... on contentBlocks_colRight_BlockType {\n                    children {\n                      ...QuestionsBlock\n                      ... on contentBlocks_group_BlockType {\n                        group: children {\n                          ... on contentBlocks_questionBlock_BlockType {\n                            questionEntries {\n                              ...QuestionEntry\n                            }\n                          }\n                        }\n                      }\n                    }\n                  }\n                }\n              }\n              ... on contentBlocks_group_BlockType {\n                group: children {\n                  ...QuestionsBlock\n                }\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
