+++ gql/public-schema/graphql.ts
@@ -21718,26 +21718,26 @@ export type ReviewPageQueryVariables = Exact<{
   >;
 }>;
 
-export type ReviewPageQuery = {
-  __typename?: "Query";
-  entry:
-    | { __typename?: "datasets_supernovaGalaxyObservations_Entry" }
-    | { __typename?: "homepage_homepage_Entry" }
-    | { __typename?: "investigations_default_Entry" }
-    | {
-        __typename?: "investigations_investigationParent_Entry";
-        title: string | null;
-      }
-    | { __typename?: "investigations_investigationSectionBreakChild_Entry" }
-    | { __typename?: "pages_pages_Entry" }
-    | { __typename?: "pages_redirectPage_Entry" }
-    | { __typename?: "questions_default_Entry" }
-    | { __typename?: "referenceModals_default_Entry" }
-    | { __typename?: "widgets_colorFilterTool_Entry" }
-    | { __typename?: "widgets_lightCurveTool_Entry" }
-    | { __typename?: "widgets_sourceSelector_Entry" }
-    | null;
-};
+
+export type ReviewPageQuery = { __typename?: 'Query', entry: { __typename?: 'datasets_supernovaGalaxyObservations_Entry' } | { __typename?: 'homepage_homepage_Entry' } | { __typename?: 'investigations_default_Entry' } | { __typename?: 'investigations_investigationParent_Entry', title: string | null, children: Array<{ __typename?: 'datasets_supernovaGalaxyObservations_Entry' } | { __typename?: 'homepage_homepage_Entry' } | { __typename?: 'investigations_default_Entry', contentBlocks: Array<{ __typename?: 'contentBlocks_barGraphTool_BlockType' } | { __typename?: 'contentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'contentBlocks_colLeft_BlockType' } | { __typename?: 'contentBlocks_colRight_BlockType' } | { __typename?: 'contentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'contentBlocks_equation_BlockType' } | { __typename?: 'contentBlocks_filterTool_BlockType' } | { __typename?: 'contentBlocks_group_BlockType', group: Array<{ __typename?: 'colorFilterToolObjects_filterimage_BlockType' } | { __typename?: 'colorFilterToolObjects_group_BlockType' } | { __typename?: 'colorFilterToolObjects_object_BlockType' } | { __typename?: 'contentBlocks_barGraphTool_BlockType' } | { __typename?: 'contentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'contentBlocks_colLeft_BlockType' } | { __typename?: 'contentBlocks_colRight_BlockType' } | { __typename?: 'contentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'contentBlocks_equation_BlockType' } | { __typename?: 'contentBlocks_filterTool_BlockType' } | { __typename?: 'contentBlocks_group_BlockType' } | { __typename?: 'contentBlocks_image_BlockType' } | (
+          { __typename?: 'contentBlocks_questionBlock_BlockType' }
+          & { ' $fragmentRefs'?: { 'QuestionsBlockFragment': QuestionsBlockFragment } }
+        ) | { __typename?: 'contentBlocks_referenceModalBlock_BlockType' } | { __typename?: 'contentBlocks_scatterplotTool_BlockType' } | { __typename?: 'contentBlocks_supernovaDistanceDistribution_BlockType' } | { __typename?: 'contentBlocks_table_BlockType' } | { __typename?: 'contentBlocks_text_BlockType' } | { __typename?: 'contentBlocks_twoColumnContainer_BlockType' } | { __typename?: 'contentBlocks_video_BlockType' } | { __typename?: 'homepageContentBlocks_image_BlockType' } | { __typename?: 'homepageContentBlocks_investigationGrid_BlockType' } | { __typename?: 'homepageContentBlocks_text_BlockType' } | { __typename?: 'multiPartBlocks_multiselect_BlockType' } | { __typename?: 'multiPartBlocks_readonlyText_BlockType' } | { __typename?: 'multiPartBlocks_select_BlockType' } | { __typename?: 'multiPartBlocks_text_BlockType' } | { __typename?: 'questionWidgetsBlock_colorFilterToolBlock_BlockType' } | { __typename?: 'questionWidgetsBlock_sourceSelectorBlock_BlockType' } | { __typename?: 'referenceContentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'referenceContentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'referenceContentBlocks_filterTool_BlockType' } | { __typename?: 'referenceContentBlocks_image_BlockType' } | { __typename?: 'referenceContentBlocks_table_BlockType' } | { __typename?: 'referenceContentBlocks_text_BlockType' } | null> | null } | { __typename?: 'contentBlocks_image_BlockType' } | (
+        { __typename?: 'contentBlocks_questionBlock_BlockType' }
+        & { ' $fragmentRefs'?: { 'QuestionsBlockFragment': QuestionsBlockFragment } }
+      ) | { __typename?: 'contentBlocks_referenceModalBlock_BlockType' } | { __typename?: 'contentBlocks_scatterplotTool_BlockType' } | { __typename?: 'contentBlocks_supernovaDistanceDistribution_BlockType' } | { __typename?: 'contentBlocks_table_BlockType' } | { __typename?: 'contentBlocks_text_BlockType' } | { __typename?: 'contentBlocks_twoColumnContainer_BlockType', columns: Array<{ __typename?: 'colorFilterToolObjects_filterimage_BlockType' } | { __typename?: 'colorFilterToolObjects_group_BlockType' } | { __typename?: 'colorFilterToolObjects_object_BlockType' } | { __typename?: 'contentBlocks_barGraphTool_BlockType' } | { __typename?: 'contentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'contentBlocks_colLeft_BlockType', children: Array<{ __typename?: 'colorFilterToolObjects_filterimage_BlockType' } | { __typename?: 'colorFilterToolObjects_group_BlockType' } | { __typename?: 'colorFilterToolObjects_object_BlockType' } | { __typename?: 'contentBlocks_barGraphTool_BlockType' } | { __typename?: 'contentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'contentBlocks_colLeft_BlockType' } | { __typename?: 'contentBlocks_colRight_BlockType' } | { __typename?: 'contentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'contentBlocks_equation_BlockType' } | { __typename?: 'contentBlocks_filterTool_BlockType' } | { __typename?: 'contentBlocks_group_BlockType', group: Array<{ __typename?: 'colorFilterToolObjects_filterimage_BlockType' } | { __typename?: 'colorFilterToolObjects_group_BlockType' } | { __typename?: 'colorFilterToolObjects_object_BlockType' } | { __typename?: 'contentBlocks_barGraphTool_BlockType' } | { __typename?: 'contentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'contentBlocks_colLeft_BlockType' } | { __typename?: 'contentBlocks_colRight_BlockType' } | { __typename?: 'contentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'contentBlocks_equation_BlockType' } | { __typename?: 'contentBlocks_filterTool_BlockType' } | { __typename?: 'contentBlocks_group_BlockType' } | { __typename?: 'contentBlocks_image_BlockType' } | (
+              { __typename?: 'contentBlocks_questionBlock_BlockType' }
+              & { ' $fragmentRefs'?: { 'QuestionsBlockFragment': QuestionsBlockFragment } }
+            ) | { __typename?: 'contentBlocks_referenceModalBlock_BlockType' } | { __typename?: 'contentBlocks_scatterplotTool_BlockType' } | { __typename?: 'contentBlocks_supernovaDistanceDistribution_BlockType' } | { __typename?: 'contentBlocks_table_BlockType' } | { __typename?: 'contentBlocks_text_BlockType' } | { __typename?: 'contentBlocks_twoColumnContainer_BlockType' } | { __typename?: 'contentBlocks_video_BlockType' } | { __typename?: 'homepageContentBlocks_image_BlockType' } | { __typename?: 'homepageContentBlocks_investigationGrid_BlockType' } | { __typename?: 'homepageContentBlocks_text_BlockType' } | { __typename?: 'multiPartBlocks_multiselect_BlockType' } | { __typename?: 'multiPartBlocks_readonlyText_BlockType' } | { __typename?: 'multiPartBlocks_select_BlockType' } | { __typename?: 'multiPartBlocks_text_BlockType' } | { __typename?: 'questionWidgetsBlock_colorFilterToolBlock_BlockType' } | { __typename?: 'questionWidgetsBlock_sourceSelectorBlock_BlockType' } | { __typename?: 'referenceContentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'referenceContentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'referenceContentBlocks_filterTool_BlockType' } | { __typename?: 'referenceContentBlocks_image_BlockType' } | { __typename?: 'referenceContentBlocks_table_BlockType' } | { __typename?: 'referenceContentBlocks_text_BlockType' } | null> | null } | { __typename?: 'contentBlocks_image_BlockType' } | (
+            { __typename?: 'contentBlocks_questionBlock_BlockType' }
+            & { ' $fragmentRefs'?: { 'QuestionsBlockFragment': QuestionsBlockFragment } }
+          ) | { __typename?: 'contentBlocks_referenceModalBlock_BlockType' } | { __typename?: 'contentBlocks_scatterplotTool_BlockType' } | { __typename?: 'contentBlocks_supernovaDistanceDistribution_BlockType' } | { __typename?: 'contentBlocks_table_BlockType' } | { __typename?: 'contentBlocks_text_BlockType' } | { __typename?: 'contentBlocks_twoColumnContainer_BlockType' } | { __typename?: 'contentBlocks_video_BlockType' } | { __typename?: 'homepageContentBlocks_image_BlockType' } | { __typename?: 'homepageContentBlocks_investigationGrid_BlockType' } | { __typename?: 'homepageContentBlocks_text_BlockType' } | { __typename?: 'multiPartBlocks_multiselect_BlockType' } | { __typename?: 'multiPartBlocks_readonlyText_BlockType' } | { __typename?: 'multiPartBlocks_select_BlockType' } | { __typename?: 'multiPartBlocks_text_BlockType' } | { __typename?: 'questionWidgetsBlock_colorFilterToolBlock_BlockType' } | { __typename?: 'questionWidgetsBlock_sourceSelectorBlock_BlockType' } | { __typename?: 'referenceContentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'referenceContentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'referenceContentBlocks_filterTool_BlockType' } | { __typename?: 'referenceContentBlocks_image_BlockType' } | { __typename?: 'referenceContentBlocks_table_BlockType' } | { __typename?: 'referenceContentBlocks_text_BlockType' } | null> | null } | { __typename?: 'contentBlocks_colRight_BlockType', children: Array<{ __typename?: 'colorFilterToolObjects_filterimage_BlockType' } | { __typename?: 'colorFilterToolObjects_group_BlockType' } | { __typename?: 'colorFilterToolObjects_object_BlockType' } | { __typename?: 'contentBlocks_barGraphTool_BlockType' } | { __typename?: 'contentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'contentBlocks_colLeft_BlockType' } | { __typename?: 'contentBlocks_colRight_BlockType' } | { __typename?: 'contentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'contentBlocks_equation_BlockType' } | { __typename?: 'contentBlocks_filterTool_BlockType' } | { __typename?: 'contentBlocks_group_BlockType', group: Array<{ __typename?: 'colorFilterToolObjects_filterimage_BlockType' } | { __typename?: 'colorFilterToolObjects_group_BlockType' } | { __typename?: 'colorFilterToolObjects_object_BlockType' } | { __typename?: 'contentBlocks_barGraphTool_BlockType' } | { __typename?: 'contentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'contentBlocks_colLeft_BlockType' } | { __typename?: 'contentBlocks_colRight_BlockType' } | { __typename?: 'contentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'contentBlocks_equation_BlockType' } | { __typename?: 'contentBlocks_filterTool_BlockType' } | { __typename?: 'contentBlocks_group_BlockType' } | { __typename?: 'contentBlocks_image_BlockType' } | { __typename?: 'contentBlocks_questionBlock_BlockType', questionEntries: Array<{ __typename?: 'datasets_supernovaGalaxyObservations_Entry' } | { __typename?: 'homepage_homepage_Entry' } | { __typename?: 'investigations_default_Entry' } | { __typename?: 'investigations_investigationParent_Entry' } | { __typename?: 'investigations_investigationSectionBreakChild_Entry' } | { __typename?: 'pages_pages_Entry' } | { __typename?: 'pages_redirectPage_Entry' } | (
+                { __typename?: 'questions_default_Entry' }
+                & { ' $fragmentRefs'?: { 'QuestionEntryFragment': QuestionEntryFragment } }
+              ) | { __typename?: 'referenceModals_default_Entry' } | { __typename?: 'widgets_colorFilterTool_Entry' } | { __typename?: 'widgets_sourceSelector_Entry' } | null> } | { __typename?: 'contentBlocks_referenceModalBlock_BlockType' } | { __typename?: 'contentBlocks_scatterplotTool_BlockType' } | { __typename?: 'contentBlocks_supernovaDistanceDistribution_BlockType' } | { __typename?: 'contentBlocks_table_BlockType' } | { __typename?: 'contentBlocks_text_BlockType' } | { __typename?: 'contentBlocks_twoColumnContainer_BlockType' } | { __typename?: 'contentBlocks_video_BlockType' } | { __typename?: 'homepageContentBlocks_image_BlockType' } | { __typename?: 'homepageContentBlocks_investigationGrid_BlockType' } | { __typename?: 'homepageContentBlocks_text_BlockType' } | { __typename?: 'multiPartBlocks_multiselect_BlockType' } | { __typename?: 'multiPartBlocks_readonlyText_BlockType' } | { __typename?: 'multiPartBlocks_select_BlockType' } | { __typename?: 'multiPartBlocks_text_BlockType' } | { __typename?: 'questionWidgetsBlock_colorFilterToolBlock_BlockType' } | { __typename?: 'questionWidgetsBlock_sourceSelectorBlock_BlockType' } | { __typename?: 'referenceContentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'referenceContentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'referenceContentBlocks_filterTool_BlockType' } | { __typename?: 'referenceContentBlocks_image_BlockType' } | { __typename?: 'referenceContentBlocks_table_BlockType' } | { __typename?: 'referenceContentBlocks_text_BlockType' } | null> | null } | { __typename?: 'contentBlocks_image_BlockType' } | (
+            { __typename?: 'contentBlocks_questionBlock_BlockType' }
+            & { ' $fragmentRefs'?: { 'QuestionsBlockFragment': QuestionsBlockFragment } }
+          ) | { __typename?: 'contentBlocks_referenceModalBlock_BlockType' } | { __typename?: 'contentBlocks_scatterplotTool_BlockType' } | { __typename?: 'contentBlocks_supernovaDistanceDistribution_BlockType' } | { __typename?: 'contentBlocks_table_BlockType' } | { __typename?: 'contentBlocks_text_BlockType' } | { __typename?: 'contentBlocks_twoColumnContainer_BlockType' } | { __typename?: 'contentBlocks_video_BlockType' } | { __typename?: 'homepageContentBlocks_image_BlockType' } | { __typename?: 'homepageContentBlocks_investigationGrid_BlockType' } | { __typename?: 'homepageContentBlocks_text_BlockType' } | { __typename?: 'multiPartBlocks_multiselect_BlockType' } | { __typename?: 'multiPartBlocks_readonlyText_BlockType' } | { __typename?: 'multiPartBlocks_select_BlockType' } | { __typename?: 'multiPartBlocks_text_BlockType' } | { __typename?: 'questionWidgetsBlock_colorFilterToolBlock_BlockType' } | { __typename?: 'questionWidgetsBlock_sourceSelectorBlock_BlockType' } | { __typename?: 'referenceContentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'referenceContentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'referenceContentBlocks_filterTool_BlockType' } | { __typename?: 'referenceContentBlocks_image_BlockType' } | { __typename?: 'referenceContentBlocks_table_BlockType' } | { __typename?: 'referenceContentBlocks_text_BlockType' } | null> | null } | { __typename?: 'contentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'contentBlocks_equation_BlockType' } | { __typename?: 'contentBlocks_filterTool_BlockType' } | { __typename?: 'contentBlocks_group_BlockType' } | { __typename?: 'contentBlocks_image_BlockType' } | { __typename?: 'contentBlocks_questionBlock_BlockType' } | { __typename?: 'contentBlocks_referenceModalBlock_BlockType' } | { __typename?: 'contentBlocks_scatterplotTool_BlockType' } | { __typename?: 'contentBlocks_supernovaDistanceDistribution_BlockType' } | { __typename?: 'contentBlocks_table_BlockType' } | { __typename?: 'contentBlocks_text_BlockType' } | { __typename?: 'contentBlocks_twoColumnContainer_BlockType' } | { __typename?: 'contentBlocks_video_BlockType' } | { __typename?: 'homepageContentBlocks_image_BlockType' } | { __typename?: 'homepageContentBlocks_investigationGrid_BlockType' } | { __typename?: 'homepageContentBlocks_text_BlockType' } | { __typename?: 'multiPartBlocks_multiselect_BlockType' } | { __typename?: 'multiPartBlocks_readonlyText_BlockType' } | { __typename?: 'multiPartBlocks_select_BlockType' } | { __typename?: 'multiPartBlocks_text_BlockType' } | { __typename?: 'questionWidgetsBlock_colorFilterToolBlock_BlockType' } | { __typename?: 'questionWidgetsBlock_sourceSelectorBlock_BlockType' } | { __typename?: 'referenceContentBlocks_cameraFilterTool_BlockType' } | { __typename?: 'referenceContentBlocks_colorFilterToolBlock_BlockType' } | { __typename?: 'referenceContentBlocks_filterTool_BlockType' } | { __typename?: 'referenceContentBlocks_image_BlockType' } | { __typename?: 'referenceContentBlocks_table_BlockType' } | { __typename?: 'referenceContentBlocks_text_BlockType' } | null> | null } | { __typename?: 'contentBlocks_video_BlockType' } | null> | null } | { __typename?: 'investigations_investigationParent_Entry' } | { __typename?: 'investigations_investigationSectionBreakChild_Entry' } | { __typename?: 'pages_pages_Entry' } | { __typename?: 'pages_redirectPage_Entry' } | { __typename?: 'questions_default_Entry' } | { __typename?: 'referenceModals_default_Entry' } | { __typename?: 'widgets_colorFilterTool_Entry' } | { __typename?: 'widgets_sourceSelector_Entry' }> } | { __typename?: 'investigations_investigationSectionBreakChild_Entry' } | { __typename?: 'pages_pages_Entry' } | { __typename?: 'pages_redirectPage_Entry' } | { __typename?: 'questions_default_Entry' } | { __typename?: 'referenceModals_default_Entry' } | { __typename?: 'widgets_colorFilterTool_Entry' } | { __typename?: 'widgets_sourceSelector_Entry' } | null };
 
 export type GlobalsQueryQueryVariables = Exact<{
   site: InputMaybe<
@@ -23800,7 +23800,7 @@ export const InvestigationChildPageMetadataDocument = {"kind":"Document","defini
 export const InvestigationMetadataDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"InvestigationMetadata"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"site"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"uri"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"entry"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"site"},"value":{"kind":"Variable","name":{"kind":"Name","value":"site"}}},{"kind":"Argument","name":{"kind":"Name","value":"uri"},"value":{"kind":"Variable","name":{"kind":"Name","value":"uri"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"investigations_investigationParent_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"title"}}]}}]}}]}}]} as unknown as DocumentNode<InvestigationMetadataQuery, InvestigationMetadataQueryVariables>;
 export const InvestigationIdDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"InvestigationId"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"site"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"uri"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"entry"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"site"},"value":{"kind":"Variable","name":{"kind":"Name","value":"site"}}},{"kind":"Argument","name":{"kind":"Name","value":"uri"},"value":{"kind":"Variable","name":{"kind":"Name","value":"uri"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"investigations_investigationParent_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"acknowledgements"},"name":{"kind":"Name","value":"text"}},{"kind":"Field","name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"uri"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"investigations_default_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"hasSavePoint"}},{"kind":"Field","name":{"kind":"Name","value":"contentBlocks"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionsBlock"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_twoColumnContainer_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"columns"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_colLeft_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionsBlock"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_group_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"group"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionsBlock"}}]}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_colRight_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionsBlock"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_group_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"group"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_questionBlock_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"Field","name":{"kind":"Name","value":"questionEntries"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionEntry"}}]}}]}}]}}]}}]}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_group_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"group"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionsBlock"}}]}}]}}]}}]}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"TableRows"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questions_default_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"rows"},"name":{"kind":"Name","value":"questionTable"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questionTable_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"cells"},"name":{"kind":"Name","value":"tableCell"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"tableCell_question_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"questionType"}},{"kind":"Field","name":{"kind":"Name","value":"options"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"options_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"optionLabel"}},{"kind":"Field","alias":{"kind":"Name","value":"value"},"name":{"kind":"Name","value":"optionValue"}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"tableCell_static_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"equation"}},{"kind":"Field","name":{"kind":"Name","value":"text"}},{"kind":"Field","name":{"kind":"Name","value":"header"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"previousQuestion"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questions_default_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"rows"},"name":{"kind":"Name","value":"questionTable"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questionTable_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"cells"},"name":{"kind":"Name","value":"tableCell"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"tableCell_question_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"questionType"}},{"kind":"Field","name":{"kind":"Name","value":"options"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"options_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"optionLabel"}},{"kind":"Field","alias":{"kind":"Name","value":"value"},"name":{"kind":"Name","value":"optionValue"}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"tableCell_static_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"equation"}},{"kind":"Field","name":{"kind":"Name","value":"text"}},{"kind":"Field","name":{"kind":"Name","value":"header"}}]}}]}}]}}]}}]}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"TabularQuestion"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questions_default_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"questionText"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"TableRows"}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"ColorFilterToolEntry"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"widgets_colorFilterTool_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"displayName"}},{"kind":"Field","name":{"kind":"Name","value":"filterToolActions"}},{"kind":"Field","alias":{"kind":"Name","value":"filterColorOptionsLabels"},"name":{"kind":"Name","value":"filterColorOptions"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"label"},"value":{"kind":"BooleanValue","value":true}}]},{"kind":"Field","alias":{"kind":"Name","value":"filterColorOptionsValues"},"name":{"kind":"Name","value":"filterColorOptions"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"label"},"value":{"kind":"BooleanValue","value":false}}]},{"kind":"Field","name":{"kind":"Name","value":"colorFilterToolObjects"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"colorFilterToolObjects_group_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"groupName"}},{"kind":"Field","alias":{"kind":"Name","value":"objects"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"colorFilterToolObjects_object_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"name"},"name":{"kind":"Name","value":"objectName"}},{"kind":"Field","alias":{"kind":"Name","value":"caption"},"name":{"kind":"Name","value":"objectCaption"}},{"kind":"Field","alias":{"kind":"Name","value":"filterImages"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"colorFilterToolObjects_filterimage_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"isEnabled"}},{"kind":"Field","name":{"kind":"Name","value":"isActive"}},{"kind":"Field","name":{"kind":"Name","value":"image"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"url"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"directUrlPreview"}}]}},{"kind":"Field","name":{"kind":"Name","value":"width"}},{"kind":"Field","name":{"kind":"Name","value":"height"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"max"},"name":{"kind":"Name","value":"colorToolMax"}},{"kind":"Field","alias":{"kind":"Name","value":"min"},"name":{"kind":"Name","value":"colorToolMin"}},{"kind":"Field","alias":{"kind":"Name","value":"defaultValue"},"name":{"kind":"Name","value":"colorToolDefaultValue"}},{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"filter"}},{"kind":"Field","alias":{"kind":"Name","value":"color"},"name":{"kind":"Name","value":"preSelectedColor"}}]}}]}}]}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"SourceSelectorEntry"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"widgets_sourceSelector_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"displayName"}},{"kind":"Field","name":{"kind":"Name","value":"dataset"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"datasets_supernovaGalaxyObservations_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"sources"},"name":{"kind":"Name","value":"alertSources"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"alertSources_source_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"color"}},{"kind":"Field","alias":{"kind":"Name","value":"x"},"name":{"kind":"Name","value":"xCoord"}},{"kind":"Field","alias":{"kind":"Name","value":"y"},"name":{"kind":"Name","value":"yCoord"}},{"kind":"Field","name":{"kind":"Name","value":"radius"}},{"kind":"Field","alias":{"kind":"Name","value":"type"},"name":{"kind":"Name","value":"sourceType"}},{"kind":"Field","alias":{"kind":"Name","value":"id"},"name":{"kind":"Name","value":"sourceName"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"galacticLongitude"}},{"kind":"Field","name":{"kind":"Name","value":"json"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"datasets_Asset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"url"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"imageAlbum"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"url"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"directUrlOriginal"}}]}},{"kind":"Field","name":{"kind":"Name","value":"width"}},{"kind":"Field","name":{"kind":"Name","value":"height"}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"QuestionEntry"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questions_default_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"answerType"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"TabularQuestion"}},{"kind":"Field","alias":{"kind":"Name","value":"options"},"name":{"kind":"Name","value":"answerOptions"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"answerOptions_option_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"optionLabel"}},{"kind":"Field","alias":{"kind":"Name","value":"value"},"name":{"kind":"Name","value":"optionValue"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"questionText"}},{"kind":"Field","name":{"kind":"Name","value":"widgetInstructions"}},{"kind":"Field","name":{"kind":"Name","value":"questionWidgetsBlock"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questionWidgetsBlock_colorFilterToolBlock_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"typeHandle"}},{"kind":"Field","name":{"kind":"Name","value":"colorFilterTool"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"ColorFilterToolEntry"}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questionWidgetsBlock_sourceSelectorBlock_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"typeHandle"}},{"kind":"Field","name":{"kind":"Name","value":"sourceSelector"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"SourceSelectorEntry"}}]}}]}}]}},{"kind":"Field","alias":{"kind":"Name","value":"parts"},"name":{"kind":"Name","value":"multiPartBlocks"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"multiPartBlocks_select_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"type"},"name":{"kind":"Name","value":"typeHandle"}},{"kind":"Field","alias":{"kind":"Name","value":"options"},"name":{"kind":"Name","value":"answerOptions"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"answerOptions_option_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"optionLabel"}},{"kind":"Field","alias":{"kind":"Name","value":"value"},"name":{"kind":"Name","value":"optionValue"}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"multiPartBlocks_text_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"type"},"name":{"kind":"Name","value":"typeHandle"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"multiPartBlocks_multiselect_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"type"},"name":{"kind":"Name","value":"typeHandle"}},{"kind":"Field","alias":{"kind":"Name","value":"options"},"name":{"kind":"Name","value":"answerOptions"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"answerOptions_option_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"optionLabel"}},{"kind":"Field","alias":{"kind":"Name","value":"value"},"name":{"kind":"Name","value":"optionValue"}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"multiPartBlocks_readonlyText_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"type"},"name":{"kind":"Name","value":"typeHandle"}},{"kind":"Field","alias":{"kind":"Name","value":"text"},"name":{"kind":"Name","value":"questionText"}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"QuestionsBlock"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_questionBlock_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"questionEntries"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionEntry"}}]}}]}}]} as unknown as DocumentNode<InvestigationIdQuery, InvestigationIdQueryVariables>;
 export const InvestigationPageDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"InvestigationPage"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"site"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"uri"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"entry"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"site"},"value":{"kind":"Variable","name":{"kind":"Name","value":"site"}}},{"kind":"Argument","name":{"kind":"Name","value":"uri"},"value":{"kind":"Variable","name":{"kind":"Name","value":"uri"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"InvestigationLandingPageTemplate"}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"InvestigationLandingPageTemplate"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"investigations_investigationParent_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"image"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"url"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"directUrlPreview"}},{"kind":"Field","name":{"kind":"Name","value":"directUrlOriginal"}},{"kind":"Field","name":{"kind":"Name","value":"PNG"}},{"kind":"Field","name":{"kind":"Name","value":"HighJPG"}},{"kind":"Field","name":{"kind":"Name","value":"LowJPG"}}]}},{"kind":"Field","name":{"kind":"Name","value":"width"}},{"kind":"Field","name":{"kind":"Name","value":"height"}},{"kind":"Field","alias":{"kind":"Name","value":"metadata"},"name":{"kind":"Name","value":"additional"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"AltTextEN"}},{"kind":"Field","name":{"kind":"Name","value":"AltTextES"}},{"kind":"Field","name":{"kind":"Name","value":"CaptionEN"}},{"kind":"Field","name":{"kind":"Name","value":"CaptionES"}},{"kind":"Field","name":{"kind":"Name","value":"Credit"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"uri"}}]}}]}}]} as unknown as DocumentNode<InvestigationPageQuery, InvestigationPageQueryVariables>;
-export const ReviewPageDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"ReviewPage"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"site"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"uri"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"entry"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"site"},"value":{"kind":"Variable","name":{"kind":"Name","value":"site"}}},{"kind":"Argument","name":{"kind":"Name","value":"uri"},"value":{"kind":"Variable","name":{"kind":"Name","value":"uri"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"investigations_investigationParent_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"title"}}]}}]}}]}}]} as unknown as DocumentNode<ReviewPageQuery, ReviewPageQueryVariables>;
+export const ReviewPageDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"ReviewPage"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"site"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"uri"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"entry"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"site"},"value":{"kind":"Variable","name":{"kind":"Name","value":"site"}}},{"kind":"Argument","name":{"kind":"Name","value":"uri"},"value":{"kind":"Variable","name":{"kind":"Name","value":"uri"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"investigations_investigationParent_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"investigations_default_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"contentBlocks"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionsBlock"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_twoColumnContainer_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"columns"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_colLeft_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionsBlock"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_group_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"group"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionsBlock"}}]}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_colRight_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionsBlock"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_group_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"group"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_questionBlock_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"questionEntries"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionEntry"}}]}}]}}]}}]}}]}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_group_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"group"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionsBlock"}}]}}]}}]}}]}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"TableRows"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questions_default_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"rows"},"name":{"kind":"Name","value":"questionTable"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questionTable_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"cells"},"name":{"kind":"Name","value":"tableCell"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"tableCell_question_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"questionType"}},{"kind":"Field","name":{"kind":"Name","value":"options"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"options_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"optionLabel"}},{"kind":"Field","alias":{"kind":"Name","value":"value"},"name":{"kind":"Name","value":"optionValue"}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"tableCell_static_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"equation"}},{"kind":"Field","name":{"kind":"Name","value":"text"}},{"kind":"Field","name":{"kind":"Name","value":"header"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"previousQuestion"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questions_default_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"rows"},"name":{"kind":"Name","value":"questionTable"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questionTable_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"cells"},"name":{"kind":"Name","value":"tableCell"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"tableCell_question_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"questionType"}},{"kind":"Field","name":{"kind":"Name","value":"options"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"options_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"optionLabel"}},{"kind":"Field","alias":{"kind":"Name","value":"value"},"name":{"kind":"Name","value":"optionValue"}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"tableCell_static_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"equation"}},{"kind":"Field","name":{"kind":"Name","value":"text"}},{"kind":"Field","name":{"kind":"Name","value":"header"}}]}}]}}]}}]}}]}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"TabularQuestion"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questions_default_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"questionText"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"TableRows"}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"ColorFilterToolEntry"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"widgets_colorFilterTool_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"displayName"}},{"kind":"Field","name":{"kind":"Name","value":"filterToolActions"}},{"kind":"Field","alias":{"kind":"Name","value":"filterColorOptionsLabels"},"name":{"kind":"Name","value":"filterColorOptions"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"label"},"value":{"kind":"BooleanValue","value":true}}]},{"kind":"Field","alias":{"kind":"Name","value":"filterColorOptionsValues"},"name":{"kind":"Name","value":"filterColorOptions"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"label"},"value":{"kind":"BooleanValue","value":false}}]},{"kind":"Field","name":{"kind":"Name","value":"colorFilterToolObjects"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"colorFilterToolObjects_group_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"groupName"}},{"kind":"Field","alias":{"kind":"Name","value":"objects"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"colorFilterToolObjects_object_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"name"},"name":{"kind":"Name","value":"objectName"}},{"kind":"Field","alias":{"kind":"Name","value":"caption"},"name":{"kind":"Name","value":"objectCaption"}},{"kind":"Field","alias":{"kind":"Name","value":"filterImages"},"name":{"kind":"Name","value":"children"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"colorFilterToolObjects_filterimage_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"isEnabled"}},{"kind":"Field","name":{"kind":"Name","value":"isActive"}},{"kind":"Field","name":{"kind":"Name","value":"image"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"url"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"directUrlPreview"}}]}},{"kind":"Field","name":{"kind":"Name","value":"width"}},{"kind":"Field","name":{"kind":"Name","value":"height"}}]}},{"kind":"Field","alias":{"kind":"Name","value":"max"},"name":{"kind":"Name","value":"colorToolMax"}},{"kind":"Field","alias":{"kind":"Name","value":"min"},"name":{"kind":"Name","value":"colorToolMin"}},{"kind":"Field","alias":{"kind":"Name","value":"defaultValue"},"name":{"kind":"Name","value":"colorToolDefaultValue"}},{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"filter"}},{"kind":"Field","alias":{"kind":"Name","value":"color"},"name":{"kind":"Name","value":"preSelectedColor"}}]}}]}}]}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"SourceSelectorEntry"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"widgets_sourceSelector_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"displayName"}},{"kind":"Field","name":{"kind":"Name","value":"dataset"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"datasets_supernovaGalaxyObservations_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"sources"},"name":{"kind":"Name","value":"alertSources"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"alertSources_source_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"color"}},{"kind":"Field","alias":{"kind":"Name","value":"x"},"name":{"kind":"Name","value":"xCoord"}},{"kind":"Field","alias":{"kind":"Name","value":"y"},"name":{"kind":"Name","value":"yCoord"}},{"kind":"Field","name":{"kind":"Name","value":"radius"}},{"kind":"Field","alias":{"kind":"Name","value":"type"},"name":{"kind":"Name","value":"sourceType"}},{"kind":"Field","alias":{"kind":"Name","value":"id"},"name":{"kind":"Name","value":"sourceName"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"galacticLongitude"}},{"kind":"Field","name":{"kind":"Name","value":"json"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"datasets_Asset"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"url"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"imageAlbum"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"url"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"directUrlOriginal"}}]}},{"kind":"Field","name":{"kind":"Name","value":"width"}},{"kind":"Field","name":{"kind":"Name","value":"height"}}]}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"QuestionEntry"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questions_default_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"answerType"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"TabularQuestion"}},{"kind":"Field","alias":{"kind":"Name","value":"options"},"name":{"kind":"Name","value":"answerOptions"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"answerOptions_option_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"optionLabel"}},{"kind":"Field","alias":{"kind":"Name","value":"value"},"name":{"kind":"Name","value":"optionValue"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"questionText"}},{"kind":"Field","name":{"kind":"Name","value":"widgetInstructions"}},{"kind":"Field","name":{"kind":"Name","value":"questionWidgetsBlock"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questionWidgetsBlock_colorFilterToolBlock_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"typeHandle"}},{"kind":"Field","name":{"kind":"Name","value":"colorFilterTool"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"ColorFilterToolEntry"}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"questionWidgetsBlock_sourceSelectorBlock_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"typeHandle"}},{"kind":"Field","name":{"kind":"Name","value":"sourceSelector"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"SourceSelectorEntry"}}]}}]}}]}},{"kind":"Field","alias":{"kind":"Name","value":"parts"},"name":{"kind":"Name","value":"multiPartBlocks"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"multiPartBlocks_select_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"type"},"name":{"kind":"Name","value":"typeHandle"}},{"kind":"Field","alias":{"kind":"Name","value":"options"},"name":{"kind":"Name","value":"answerOptions"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"answerOptions_option_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"optionLabel"}},{"kind":"Field","alias":{"kind":"Name","value":"value"},"name":{"kind":"Name","value":"optionValue"}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"multiPartBlocks_text_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"type"},"name":{"kind":"Name","value":"typeHandle"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"multiPartBlocks_multiselect_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"type"},"name":{"kind":"Name","value":"typeHandle"}},{"kind":"Field","alias":{"kind":"Name","value":"options"},"name":{"kind":"Name","value":"answerOptions"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"answerOptions_option_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"label"},"name":{"kind":"Name","value":"optionLabel"}},{"kind":"Field","alias":{"kind":"Name","value":"value"},"name":{"kind":"Name","value":"optionValue"}}]}}]}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"multiPartBlocks_readonlyText_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","alias":{"kind":"Name","value":"type"},"name":{"kind":"Name","value":"typeHandle"}},{"kind":"Field","alias":{"kind":"Name","value":"text"},"name":{"kind":"Name","value":"questionText"}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"QuestionsBlock"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"contentBlocks_questionBlock_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"questionEntries"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"QuestionEntry"}}]}}]}}]} as unknown as DocumentNode<ReviewPageQuery, ReviewPageQueryVariables>;
 export const GlobalsQueryDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"GlobalsQuery"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"site"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"siteInfo"},"name":{"kind":"Name","value":"globalSet"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"site"},"value":{"kind":"Variable","name":{"kind":"Name","value":"site"}}},{"kind":"Argument","name":{"kind":"Name","value":"handle"},"value":{"kind":"StringValue","value":"siteInfo","block":false}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"siteInfo_GlobalSet"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"language"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"handle"}},{"kind":"Field","name":{"kind":"Name","value":"siteTitle"}},{"kind":"Field","name":{"kind":"Name","value":"siteDescription"}}]}}]}},{"kind":"Field","alias":{"kind":"Name","value":"menuContent"},"name":{"kind":"Name","value":"globalSet"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"site"},"value":{"kind":"Variable","name":{"kind":"Name","value":"site"}}},{"kind":"Argument","name":{"kind":"Name","value":"handle"},"value":{"kind":"StringValue","value":"menuContent","block":false}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"menuContent_GlobalSet"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"helpUrl"}}]}}]}}]}}]} as unknown as DocumentNode<GlobalsQueryQuery, GlobalsQueryQueryVariables>;
 export const HomepageQueryDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"HomepageQuery"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"site"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"uri"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"entry"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"site"},"value":{"kind":"Variable","name":{"kind":"Name","value":"site"}}},{"kind":"Argument","name":{"kind":"Name","value":"uri"},"value":{"kind":"Variable","name":{"kind":"Name","value":"uri"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"HomepageTemplate"}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"InvestigationGridBlock"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"homepageContentBlocks_investigationGrid_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","alias":{"kind":"Name","value":"heading"},"name":{"kind":"Name","value":"contentHeading"}},{"kind":"Field","alias":{"kind":"Name","value":"investigations"},"name":{"kind":"Name","value":"investigationEntries"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"investigations_investigationParent_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","alias":{"kind":"Name","value":"rawImage"},"name":{"kind":"Name","value":"image"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"url"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"directUrlPreview"}},{"kind":"Field","name":{"kind":"Name","value":"directUrlOriginal"}},{"kind":"Field","name":{"kind":"Name","value":"PNG"}},{"kind":"Field","name":{"kind":"Name","value":"HighJPG"}},{"kind":"Field","name":{"kind":"Name","value":"LowJPG"}}]}},{"kind":"Field","name":{"kind":"Name","value":"width"}},{"kind":"Field","name":{"kind":"Name","value":"height"}},{"kind":"Field","alias":{"kind":"Name","value":"metadata"},"name":{"kind":"Name","value":"additional"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"AltTextEN"}},{"kind":"Field","name":{"kind":"Name","value":"AltTextES"}},{"kind":"Field","name":{"kind":"Name","value":"CaptionEN"}},{"kind":"Field","name":{"kind":"Name","value":"CaptionES"}},{"kind":"Field","name":{"kind":"Name","value":"Credit"}}]}}]}},{"kind":"Field","name":{"kind":"Name","value":"url"}}]}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"HomepageTemplate"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"homepage_homepage_Entry"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","alias":{"kind":"Name","value":"contentBlocks"},"name":{"kind":"Name","value":"homepageContentBlocks"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"homepageContentBlocks_text_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"text"}}]}},{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"homepageContentBlocks_image_BlockType"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"caption"}},{"kind":"Field","name":{"kind":"Name","value":"layout"}},{"kind":"Field","name":{"kind":"Name","value":"image"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"url"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"directUrlPreview"}},{"kind":"Field","name":{"kind":"Name","value":"directUrlOriginal"}},{"kind":"Field","name":{"kind":"Name","value":"PNG"}},{"kind":"Field","name":{"kind":"Name","value":"HighJPG"}},{"kind":"Field","name":{"kind":"Name","value":"LowJPG"}},{"kind":"Field","name":{"kind":"Name","value":"preview"}}]}},{"kind":"Field","name":{"kind":"Name","value":"width"}},{"kind":"Field","name":{"kind":"Name","value":"height"}},{"kind":"Field","alias":{"kind":"Name","value":"metadata"},"name":{"kind":"Name","value":"additional"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"AltTextEN"}},{"kind":"Field","name":{"kind":"Name","value":"AltTextES"}},{"kind":"Field","name":{"kind":"Name","value":"CaptionEN"}},{"kind":"Field","name":{"kind":"Name","value":"CaptionES"}},{"kind":"Field","name":{"kind":"Name","value":"Credit"}}]}}]}}]}},{"kind":"FragmentSpread","name":{"kind":"Name","value":"InvestigationGridBlock"}}]}}]}}]} as unknown as DocumentNode<HomepageQueryQuery, HomepageQueryQueryVariables>;
 export const PagePreviewQueryDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"PagePreviewQuery"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"site"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"uri"}},"type":{"kind":"ListType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"entry"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"site"},"value":{"kind":"Variable","name":{"kind":"Name","value":"site"}}},{"kind":"Argument","name":{"kind":"Name","value":"uri"},"value":{"kind":"Variable","name":{"kind":"Name","value":"uri"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"Field","name":{"kind":"Name","value":"uri"}},{"kind":"Field","name":{"kind":"Name","value":"title"}}]}}]}}]} as unknown as DocumentNode<PagePreviewQueryQuery, PagePreviewQueryQueryVariables>;
